# docs/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "O-Week v1.1 API",
    "version": "1.1.0",
    "description": "Formal User-based API. Authentication uses the HttpOnly owk_session cookie."
  },
  "servers": [{ "url": "/api/v1" }],
...
  "paths": {
...
    "/admin/accounts": { "get": { "tags": ["Admin"], "operationId": "listAdminAccounts", "security": [{ "sessionCookie": [] }], "description": "Lists accounts with ACTIVE accounts before ARCHIVED accounts by default. Set status=ACTIVE to hide archived accounts from the result.", "parameters": [{ "name": "status", "in": "query", "schema": { "type": "string", "enum": ["ACTIVE", "ARCHIVED"] } }], "responses": { "200": { "$ref": "#/components/responses/Success" } } } },
...
    "/admin/accounts/import": { "post": { "tags": ["Admin"], "operationId": "importAdminAccounts", "security": [{ "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }], "responses": { "200": { "$ref": "#/components/responses/Success" }, "409": { "$ref": "#/components/responses/Error" } } } },
...
    "/admin/accounts/{id}": { "patch": { "tags": ["Admin"], "operationId": "updateAdminAccount", "security": [{ "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/UserId" }, { "$ref": "#/components/parameters/IdempotencyKey" }], "responses": { "200": { "$ref": "#/components/responses/Success" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } } },
...
    "/admin/audit-logs": { "get": { "tags": ["Admin"], "operationId": "listAdminAuditLogs", "security": [{ "sessionCookie": [] }], "parameters": [{ "name": "cursor", "in": "query", "schema": { "type": "string" } }, { "name": "from", "in": "query", "schema": { "type": "string", "format": "date-time" } }, { "name": "to", "in": "query", "schema": { "type": "string", "format": "date-time" } }], "responses": { "200": { "$ref": "#/components/responses/Success" } } } }
  },
  "components": {
    "securitySchemes": { "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "owk_session" } },
    "parameters": {
...
      "UserId": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
...
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 200 } }
    },
    "schemas": {
...
      "ErrorEnvelope": { "type": "object", "required": ["error", "requestId"], "properties": { "error": { "type": "object", "required": ["code", "message", "details"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "details": { "type": "object" } } }, "requestId": { "type": "string", "pattern": "^req_" } } },
      "SuccessEnvelope": { "type": "object", "required": ["data", "requestId"], "properties": { "data": {}, "requestId": { "type": "string", "pattern": "^req_" } } }
    },
    "responses": {
      "Success": { "description": "Successful response", "headers": { "Cache-Control": { "schema": { "type": "string" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessEnvelope" } } } },
      "Error": { "description": "Stable error envelope", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
    }
  }
}